Some refactoring - getting back into this project - Re-organized Items namespace into Misc classes & actual Item classes - Allow Weapon to rolls its own damage
