+++ RPG_Game/RPG_Game_Engine/Factories/ItemFactory.cs
@@ -3,7 +3,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Collections.Generic;
-using Engine.Models.Items;
+using Engine.Models.Items.Item;
+using Engine.Models.Items.Misc;
 
 namespace Engine.Factories
 {
