+++ RPG_Game/RPG_Game_Engine/Models/Inventory.cs
@@ -3,7 +3,8 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Collections.ObjectModel;
-using Engine.Models.Items;
+using Engine.Models.Items.Item;
+using Engine.Models.Items.Misc;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using Engine.Factories;
